refactor: Move parameter resolver tests
